Explicit boolean check on single module check
I tried running Kansa on a Windows XP SP3 w/ PowerShell 2.0 and couldn't get past this check. Using an explicit boolean comparison made it work. I'm not sure if this was a language feature that changed. I wanted to use Kansa to help evaluate malware indicators of compromise on known evil code.