Get-Procdump.ps1
With Kansa's refactoring, all modules that return binary data have to be
refactored as well so they return objects, with binary data as a
property. Get-Procdump.ps1 now returns a base64 encoded GZip stream that
contains the process memory dump. Currently this script writes two files
to disk, so I'm not pleased with it. I need to see if I can read the
first dmp file into a memory stream and then compress and base64 encode
that stream in memory, but for now, this gets memory dump data, however,
ugly it may be.