Get-Procdump.ps1 compressed memory
This took far too long to figure out, but it reads the dmp file from
disk, then compresses it in memory, then base 64 encodes the memory
stream and sticks that result in an object property and returns the
object. Kansa will write the data to disk in the format specified at the
command line (e.g. csv (default), tsv, xml).